pyproject.toml
[tool.poetry]
name = "prbot-telegram"
version = "0.1.0"
description = ""
authors = ["rg.batyrshin <rg.batyrshin@localhost>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.2.0"
asyncpg = "^0.29.0"
sqlalchemy = "1.4.50"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
redis = "^5.0.1"
aioredis = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
